add order completed

+++ repository/cart.go
@@ -2,6 +2,8 @@ package repository
 
 import (
 	//"fmt"
+
+	"fmt"
 	"github/jaseelaali/orchid/database"
 	"github/jaseelaali/orchid/models"
 )
@@ -14,12 +16,12 @@ func Addcart(P_Id, Quantity, U_Id int) error {
 	//fmt.Println("************* %v"productname)
 
 	err = database.DB.Create(&models.CartItem{
-		User_id:       U_Id,
-		Product_Name:  productname,
-		Product_Id:    P_Id,
-		Quantity:      Quantity,
-		Product_Price: productprice,
-		Total_price:   (productprice * Quantity),
+		User_id:             U_Id,
+		Product_Name:        productname,
+		Product_Id:          P_Id,
+		Quantity:            Quantity,
+		Product_Price:       productprice,
+		Product_Total_price: (productprice * Quantity),
 	}).Where("user_id", U_Id)
 	if err != nil {
 		return err.Error
@@ -37,12 +39,13 @@ func ADDcart(newquantity, P_Id, Quantity, U_Id int) error {
 	price := ((newquantity + Quantity) * productprice)
 	//fmt.Println("****pazhe:%v****puth:%v***quq:%v", newquantity, Quantity, quantity)
 	err = database.DB.Raw("UPDATE cart_items SET quantity=$1 WHERE user_id=$2 AND product_id=$3;", quantity, U_Id, P_Id).Scan(&models.CartItem{})
-	err = database.DB.Raw("UPDATE cart_items SET total_price=$1 WHERE product_id=$2  AND user_id=$3;", price, P_Id, U_Id).Scan(&models.CartItem{})
+	err = database.DB.Raw("UPDATE cart_items SET product_total_price=$1 WHERE product_id=$2  AND user_id=$3;", price, P_Id, U_Id).Scan(&models.CartItem{})
 	if err != nil {
 		return err.Error
 	}
 	return nil
 }
+
 func Viewcart(ID int) ([]models.CartItem, error) {
 	Id := ID
 	//fmt.Println("************* %v", Id)
@@ -62,18 +65,47 @@ func Deleteitem(id, QUantity, User_Id int) error {
 	var OldQuantity, Price int
 	err := database.DB.Raw("SELECT quantity FROM cart_items WHERE user_id=$1 AND product_id=$2;", U_Id, Id).Scan(&OldQuantity)
 	err = database.DB.Raw("SELECT product_price FROM cart_items WHERE user_id=$1 AND product_id=$2;", U_Id, Id).Scan(&Price)
+	if OldQuantity >= 0 {
 
-	quantity := (OldQuantity - Quantity)
-	price := Price * quantity
-	//fmt.Println("****old:%v**kal:%v**new:%v***price:%v", OldQuantity, Quantity, quantity, price)
-	err = database.DB.Raw("UPDATE cart_items SET quantity=$1 WHERE user_id=$2 AND product_id=$3;", quantity, U_Id, Id).Scan(&models.CartItem{})
-	err = database.DB.Raw("UPDATE cart_items SET total_price=$1 WHERE user_id=$2 AND product_id=$3;", price, U_Id, Id).Scan(&models.CartItem{})
+		quantity := (OldQuantity - Quantity)
+		price := Price * quantity
+		//fmt.Println("****old:%v**kal:%v**new:%v***price:%v", OldQuantity, Quantity, quantity, price)
+		err = database.DB.Raw("UPDATE cart_items SET quantity=$1 WHERE user_id=$2 AND product_id=$3;", quantity, U_Id, Id).Scan(&models.CartItem{})
+		err = database.DB.Raw("UPDATE cart_items SET product_total_price=$1 WHERE user_id=$2 AND product_id=$3;", price, U_Id, Id).Scan(&models.CartItem{})
 
-	if quantity == 0 {
-		database.DB.Raw("DELETE FROM cart_items WHERE user_id=$1 AND product_id=$2;", U_Id, Id).Scan(&models.CartItem{})
-	}
-	if err.Error != nil {
+		if quantity == 0 {
+			database.DB.Raw("DELETE FROM cart_items WHERE user_id=$1 AND product_id=$2;", U_Id, Id).Scan(&models.CartItem{})
+		}
+		if err.Error != nil {
+			return err.Error
+		}
+		return nil
+	} else {
 		return err.Error
 	}
-	return nil
+}
+
+/*
+	func ADCart(product_id, user_id int) error {
+		result := database.DB.Raw("SELECT * from carts where user_id=$1;", user_id).Scan(&models.Cart{})
+		if result.Error != nil {
+			//cart_items := &models.CartItem{}
+			var price int
+			result = database.DB.Raw("SELECT product_total_price from cart_items where user_id=$1 AND product_id=$2;", user_id, product_id).Scan(&price)
+			fmt.Println("******")
+			fmt.Println("\\\\\\\\\\\\", price)
+		}
+		return nil
+	}
+*/
+
+func SumCart(user_id int) (int, error) {
+
+	var sum int
+	result := database.DB.Raw("SELECT SUM(product_total_price)FROM cart_items WHERE user_id=$1;", user_id).Scan(&sum)
+	fmt.Println("\n\nthe sum is :", sum, user_id)
+	if result != nil {
+		return sum, result.Error
+	}
+	return sum, nil
 }
